Switch to invoking server directly
