+++ README.md
@@ -3,4 +3,5 @@ Utilities for evaluating tsinfer inferences
 
 
 To launch the app use:
-`panel serve app.py --args /path/to/trees-file`
+
+`python main.py /path/to/trees-file`
